backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.2
PyJWT==2.8.0
python-multipart==0.0.9
